requirements.txt
streamlit
pandas
numpy
scikit-learn
spotipy
python-dotenv
pyarrow


